+++ crates/swc_ecma_fast_parser/src/lexer/jsx.rs
@@ -107,7 +107,7 @@ impl Lexer<'_> {
     /// Read JSX text content
     fn read_jsx_text(&mut self, had_line_break: bool) -> Result<Token> {
         let start_pos = self.start_pos;
-        let start_idx = start_pos.0 as usize;
+        let start_idx = start_pos.0;
 
         let mut text = String::new();
